requirement.txt
Flask==2.2.2
psycopg2-binary==2.9.3
SQLAlchemy==1.4.41
jwt==1.3.1
PyJWT==2.6.0
uuid==1.30
google-cloud-storage==2.1.0
Flask-Cors==3.0.10 
python-dotenv==0.21.0  
pandas==1.5.2